Rename and move acquire process from customer to account.
